requirements.txt
# Komplete Kontrol libs
hidapi

# Build libs
cx_Freeze
